TR-1436 Moved template containers alongside page components
