+++ src/pages/templates/index.js
@@ -1,10 +1,10 @@
-import CreatePage from './containers/CreatePage.container';
-import EditPage from './containers/EditPage.container';
-import EditAndPreviewPage from './containers/EditAndPreviewPage.container';
-import ListPage from './containers/ListPage.container';
-import PreviewDraftPage from './containers/PreviewDraftPage.container';
-import PreviewPublishedPage from './containers/PreviewPublishedPage.container';
-import PublishedPage from './containers/PublishedPage.container';
+import CreatePage from './CreatePage.container';
+import EditPage from './EditPage.container';
+import EditAndPreviewPage from './EditAndPreviewPage.container';
+import ListPage from './ListPage.container';
+import PreviewDraftPage from './PreviewDraftPage.container';
+import PreviewPublishedPage from './PreviewPublishedPage.container';
+import PublishedPage from './PublishedPage.container';
 
 export default {
   CreatePage,
